Refactor to fixtures
