+++ spec/frontend/layout.spec.js
@@ -1,15 +1,8 @@
-import { test, expect } from '@playwright/test';
+import { test, expect } from './fixtures.js'
 
 test.describe("vhosts", _ => {
-  const vhosts = ['foo', 'bar']
-
-  test.beforeEach(async ({ page }) => {
-    const vhostResponse = vhosts.map(x => { return {name: x} })
-    await page.route(/\/api\/vhosts/, route => route.fulfill({ json: vhostResponse }))
+  test('are loaded', async ({ page, baseURL, vhosts }) => {
     await page.goto('/')
-  })
-
-  test('are loaded', async ({ page }) => {
     const vhostOptions = await page.locator('#userMenuVhost option').allTextContents()
     vhosts.forEach(vhost => {
       expect(vhostOptions).toContain(vhost)
@@ -17,6 +10,7 @@ test.describe("vhosts", _ => {
   })
 
   test('remember selection', async ({ page }) => {
+    await page.goto('/')
     await page.locator('#userMenuVhost').selectOption('foo')
     await expect(page.locator('#userMenuVhost option:checked')).toHaveText(['foo'])
   })
